fixup go lint and Go version compatibility
