+++ aws/session/session_test.go
@@ -406,11 +406,3 @@ func TestNewSessionWithOptions_Overrides(t *testing.T) {
 		})
 	}
 }
-
-func initSessionTestEnv() (oldEnv []string) {
-	oldEnv = awstesting.StashEnv()
-	os.Setenv("AWS_CONFIG_FILE", "file_not_exists")
-	os.Setenv("AWS_SHARED_CREDENTIALS_FILE", "file_not_exists")
-
-	return oldEnv
-}
