+++ aws/session/shared_test.go
@@ -1,0 +1,15 @@
+package session
+
+import (
+	"os"
+
+	"github.com/aws/aws-sdk-go/awstesting"
+)
+
+func initSessionTestEnv() (oldEnv []string) {
+	oldEnv = awstesting.StashEnv()
+	os.Setenv("AWS_CONFIG_FILE", "file_not_exists")
+	os.Setenv("AWS_SHARED_CREDENTIALS_FILE", "file_not_exists")
+
+	return oldEnv
+}
